+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/core/KruskalMotifPruner.java
@@ -2,7 +2,6 @@ package org.rcsb.strucmotif.core;
 
 import org.rcsb.strucmotif.domain.motif.ResiduePairOccurrence;
 import org.rcsb.strucmotif.domain.structure.ResidueGraph;
-import org.rcsb.strucmotif.domain.structure.ResidueType;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -21,9 +20,9 @@ public class KruskalMotifPruner implements MotifPruner {
     }
 
     @Override
-    public List<ResiduePairOccurrence> prune(ResidueGraph residueGraph, Map<Integer, Set<ResidueType>> exchanges) {
+    public List<ResiduePairOccurrence> prune(ResidueGraph residueGraph) {
         List<ResiduePairOccurrence> original = residueGraph.residuePairOccurrencesSequential().collect(Collectors.toList());
-        List<ResiduePairOccurrence> sorted = ResiduePairOccurrence.sort(original, exchanges);
+        List<ResiduePairOccurrence> sorted = ResiduePairOccurrence.sort(original);
 
         // ignore motifs with <4 identifiers
         if (residueGraph.getResidueCount() < 4) {