+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/core/NoOperationMotifPruner.java
@@ -2,12 +2,9 @@ package org.rcsb.strucmotif.core;
 
 import org.rcsb.strucmotif.domain.motif.ResiduePairOccurrence;
 import org.rcsb.strucmotif.domain.structure.ResidueGraph;
-import org.rcsb.strucmotif.domain.structure.ResidueType;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
@@ -22,8 +19,8 @@ public class NoOperationMotifPruner implements MotifPruner {
     }
 
     @Override
-    public List<ResiduePairOccurrence> prune(ResidueGraph residueGraph, Map<Integer, Set<ResidueType>> exchanges) {
+    public List<ResiduePairOccurrence> prune(ResidueGraph residueGraph) {
         List<ResiduePairOccurrence> original = residueGraph.residuePairOccurrencesSequential().collect(Collectors.toList());
-        return ResiduePairOccurrence.sort(original, exchanges);
+        return ResiduePairOccurrence.sort(original);
     }
 }