+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/domain/query/StructureQuery.java
@@ -13,7 +13,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * The immutable container for a structural motif query.
@@ -57,10 +56,7 @@ public class StructureQuery implements SearchQuery<StructureParameters, Structur
             throw new IllegalQueryDefinitionException("Query violates distance threshold");
         }
 
-        Map<Integer, Set<ResidueType>> mappedExchanges = exchanges.entrySet()
-                .stream()
-                .collect(Collectors.toMap(e -> structure.getResidueIndex(e.getKey()), Map.Entry::getValue));
-        List<ResiduePairOccurrence> residuePairOccurrences = parameters.getMotifPruner().prune(residueGraph, mappedExchanges);
+        List<ResiduePairOccurrence> residuePairOccurrences = parameters.getMotifPruner().prune(residueGraph);
         this.queryStructure = new StructureQueryStructure(structureIdentifier, structure, labelSelections, residues, residuePairOccurrences, exchanges);
         this.parameters = parameters;
         this.exchanges = exchanges;