+++ strucmotif-search-core/src/test/java/org/rcsb/strucmotif/core/StructureIntegrationTest.java
@@ -160,7 +160,6 @@ class StructureIntegrationTest {
                 .backboneDistanceTolerance(1)
                 .sideChainDistanceTolerance(1)
                 .angleTolerance(1)
-                .rmsdCutoff(3.0) // need higher value to see swaps
                 .buildParameters()
                 .addPositionSpecificExchange(new LabelSelection("A", "1", 162), Set.of(ResidueType.LYSINE, ResidueType.HISTIDINE))
                 .addPositionSpecificExchange(new LabelSelection("A", "1", 245), Set.of(ResidueType.GLUTAMIC_ACID, ResidueType.ASPARTIC_ACID, ResidueType.ASPARAGINE))
@@ -168,7 +167,7 @@ class StructureIntegrationTest {
 
         StructureSearchResult response = buildParameters.buildContext().run();
 
-        assertEquals(891, response.getHits().size());
+        assertEquals(788, response.getHits().size());
 
         List<String> observedExchanges = response.getHits()
                 .stream()